➖ Simplify dependencies & instructions
Remove extra tools (just, uv) from instructions --
they are still there if you want to use them
but not mentioned so as to keep the demo as simple as possible.

Remove valohai-cli from the requirements --
it is not needed to run the execution.
It can be installed in the venv if the user
wants to use the CLI
but no need to always install it.